+++ src/Core/Tools/ObjectUpdate.php
@@ -94,7 +94,7 @@ class ObjectUpdate extends AbstractMCPTool
 					'description' => 'The ID of the object to update.',
 				],
 				'fields' => [
-					'type'                 => 'object',
+					'type'                 => MCPHelper::MAP_TYPE,
 					'description'          => 'Key/value pairs to update. Keys are attribute codes. Call core_class_schema for the attribute codes of the class, their types and which ones are mandatory.',
 					'additionalProperties' => true,
 				],